// src/app/models/response/LookupResponse.ts
export class LookupResponse {
  id?: string;
  uuid?: string;
  lookupEnumKey?: string;
  lookupEnumValue?: string;
  lookupName?: string;
  lookupShortName?: string;
  description?: string;
  image?: string;
  parentId?: string;
  parentName?: string;
  parentFullName?: string;
  enabled?: boolean;
}